bitfinex_longs/requirements.txt
requests
pandas
yfinance
matplotlib
python-dotenv